huggingface-deployment/app.py
import os
import gradio as gr
import numpy as np
import cv2
from PIL import Image, ImageDraw, ImageFilter
import torch
import torch.nn as nn
import torch.nn.functional as F
from torchvision import transforms
import requests
from io import BytesIO
import base64

# 設定日誌
import logging
logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

class FaceDancerInference:
    def __init__(self):
        self.device = torch.device('cuda' if torch.cuda.is_available() else 'cpu')
        self.transform = transforms.Compose([
            transforms.Resize((256, 256)),
            transforms.ToTensor(),
            transforms.Normalize(mean=[0.5, 0.5, 0.5], std=[0.5, 0.5, 0.5])
        ])
        
    def preprocess_image(self, image):
        """預處理圖片"""
        if isinstance(image, str):
            # 如果是 URL，下載圖片
            response = requests.get(image)
            image = Image.open(BytesIO(response.content))
        elif isinstance(image, np.ndarray):
            image = Image.fromarray(image)
        
        # 轉換為 RGB
        if image.mode != 'RGB':
            image = image.convert('RGB')
        
        # 調整大小
        image = image.resize((256, 256))
        
        return image
    
    def detect_face(self, image):
        """簡單的臉部檢測（中心裁剪）"""
        # 這裡應該使用更複雜的臉部檢測算法
        # 為了簡化，我們使用中心裁剪
        width, height = image.size
        size = min(width, height)
        left = (width - size) // 2
        top = (height - size) // 2
        right = left + size
        bottom = top + size
        
        face = image.crop((left, top, right, bottom))
        return face
    
    def simulate_face_swap(self, source_image, target_image):
        """智能臉部交換模擬"""
        try:
            # 預處理圖片
            source_face = self.preprocess_image(source_image)
            target_img = self.preprocess_image(target_image)
            
            # 獲取圖片尺寸
            target_width, target_height = target_img.size
            
            # 計算臉部區域
            face_size = min(target_width, target_height) * 0.4
            face_x = (target_width - face_size) // 2
            face_y = int(target_height * 0.15)
            
            # 調整源臉部大小
            source_face = source_face.resize((int(face_size), int(face_size)), Image.LANCZOS)
            
            # 創建橢圓形遮罩
            mask = Image.new('L', (int(face_size), int(face_size)), 0)
            mask_draw = ImageDraw.Draw(mask)
            
            # 繪製橢圓形遮罩
            ellipse_bbox = [0, 0, face_size, face_size]
            mask_draw.ellipse(ellipse_bbox, fill=255)
            
            # 應用高斯模糊到遮罩邊緣
            mask = mask.filter(ImageFilter.GaussianBlur(radius=5))
            
            # 合成臉部
            target_img.paste(source_face, (face_x, face_y), mask)
            
            # 顏色調整
            target_array = np.array(target_img)
            
            # 簡單的顏色平衡
            target_array = target_array.astype(np.float32)
            target_array = np.clip(target_array * 1.05, 0, 255).astype(np.uint8)
            
            result = Image.fromarray(target_array)
            
            return result
            
        except Exception as e:
            logger.error(f"臉部交換失敗: {e}")
            return target_image
    
    def perform_face_swap(self, source_image, target_image):
        """執行臉部交換"""
        try:
            logger.info("開始臉部交換...")
            
            # 使用模擬的臉部交換
            result = self.simulate_face_swap(source_image, target_image)
            
            logger.info("臉部交換完成")
            return result
            
        except Exception as e:
            logger.error(f"臉部交換錯誤: {e}")
            return target_image

# 初始化推理引擎
inference_engine = FaceDancerInference()

def face_swap_interface(source_image, target_image):
    """Gradio 介面函數"""
    try:
        if source_image is None or target_image is None:
            return None, "請上傳源臉部和目標圖片"
        
        # 執行臉部交換
        result = inference_engine.perform_face_swap(source_image, target_image)
        
        return result, "臉部交換完成！"
        
    except Exception as e:
        logger.error(f"介面錯誤: {e}")
        return None, f"錯誤: {str(e)}"

# 建立 Gradio 介面
with gr.Blocks(title="FaceDancer - AI 臉部交換", theme=gr.themes.Soft()) as demo:
    gr.Markdown("# 🎭 FaceDancer - AI 臉部交換")
    gr.Markdown("上傳源臉部圖片和目標圖片，AI 將自動進行臉部交換")
    
    with gr.Row():
        with gr.Column():
            gr.Markdown("### 📸 源臉部圖片")
            source_input = gr.Image(label="上傳源臉部圖片", type="pil")
            
        with gr.Column():
            gr.Markdown("### 🎯 目標圖片")
            target_input = gr.Image(label="上傳目標圖片", type="pil")
    
    with gr.Row():
        swap_button = gr.Button("🎭 開始臉部交換", variant="primary", size="lg")
    
    with gr.Row():
        with gr.Column():
            gr.Markdown("### 🎨 交換結果")
            result_output = gr.Image(label="臉部交換結果", type="pil")
        
        with gr.Column():
            gr.Markdown("### 📝 狀態")
            status_output = gr.Textbox(label="處理狀態", interactive=False)
    
    # 事件處理
    swap_button.click(
        fn=face_swap_interface,
        inputs=[source_input, target_input],
        outputs=[result_output, status_output]
    )
    
    # 範例圖片
    gr.Markdown("### 📋 使用說明")
    gr.Markdown("""
    1. **上傳源臉部圖片**: 包含要交換的臉部
    2. **上傳目標圖片**: 要進行臉部交換的目標圖片
    3. **點擊開始臉部交換**: AI 將自動處理
    4. **查看結果**: 在右側查看臉部交換結果
    
    **注意事項**:
    - 確保臉部圖片清晰可見
    - 建議使用正面臉部照片
    - 處理時間約 10-30 秒
    """)

# 啟動應用
if __name__ == "__main__":
    demo.launch(server_name="0.0.0.0", server_port=7860)
