huggingface-deployment/requirements.txt
gradio>=4.0.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0.0
numpy>=1.21.0
opencv-python>=4.5.0
requests>=2.25.0
transformers>=4.20.0
accelerate>=0.20.0
fastapi>=0.100.0
uvicorn>=0.20.0
python-multipart>=0.0.6
